Main/Unattended: Try use the new UDF re-mastering for unattended Windows installations (Vista and later). [build fix] bugref:10517 jiraref:VBP-1922
svn:sync-xref-src-repo-rev: r171005

+++ src/VBox/Main/src-server/UnattendedInstaller.cpp
@@ -1,4 +1,4 @@
-/* $Id: UnattendedInstaller.cpp 111444 2025-10-18 18:40:47Z knut.osmundsen@oracle.com $ */
+/* $Id: UnattendedInstaller.cpp 111446 2025-10-18 23:11:22Z knut.osmundsen@oracle.com $ */
 /** @file
  * UnattendedInstaller class and it's descendants implementation
  */
@@ -84,7 +84,7 @@ static uint64_t hlpVfsGetFileSize(RTVFS hVfs, const char *pszPath)
 {
     RTFSOBJINFO ObjInfo;
     int vrc = RTVfsQueryPathInfo(hVfs, pszPath, &ObjInfo, RTFSOBJATTRADD_NOTHING, RTPATH_F_FOLLOW_LINK);
-    return RT_SUCCESS(vrc) && RTFS_IS_FILE(ObjInfo.Attr.fMode) ? ObjInfo.cbObject : UINT64_MAX;
+    return RT_SUCCESS(vrc) && RTFS_IS_FILE(ObjInfo.Attr.fMode) ? (uint64_t)ObjInfo.cbObject : UINT64_MAX;
 }
